Compete test setup
