+++ client/index.js
@@ -1,27 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      reviews: []
-    };
-  }
-
-  componentDidMount() {
-    fetch('http://localhost:3000/reviews/1')
-      // .then(response => response.json())
-      .then(response => console.log(response.json()))
-      // .then(reviews => this.setState({ reviews }))
-  }
-
-  render() {
-    return (
-      <div>{this.state.reviews}</div>
-    );
-  }
-}
+import App from './components/App.jsx';
 
 ReactDOM.render(<App />, document.getElementById('root'));
