+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "",
+  "name": "Reviews",
   "version": "",
   "description": "",
-  "author": "",
+  "author": "Grant Taggart",
   "license": "",
   "engines": {
     "node": ">=6.13.0"
@@ -11,12 +11,18 @@
     "start": "nodemon server/index.js",
     "watch": "webpack serve --open",
     "build": "webpack",
-    "seed": "node db/seeder.js"
+    "seed": "node db/seeder.js",
+    "test": "jest --coverage"
   },
   "dependencies": {
+    "@babel/plugin-transform-runtime": "^7.13.10",
+    "@babel/runtime": "^7.13.10",
     "body-parser": "^1.19.0",
+    "chai": "^4.3.4",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.6",
     "express": "^4.17.1",
     "faker": "^5.4.0",
     "http-errors": "^1.8.0",
@@ -26,14 +32,15 @@
     "superagent": "^6.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.8",
-    "@babel/preset-env": "^7.13.9",
+    "@babel/core": "^7.13.10",
+    "@babel/preset-env": "^7.13.10",
     "@babel/preset-react": "^7.12.13",
     "babel-loader": "^8.2.2",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "html-webpack-plugin": "^5.2.0",
+    "jest": "^26.6.3",
     "nodemon": "^2.0.7",
-    "webpack": "^5.24.3",
+    "webpack": "^5.25.0",
     "webpack-cli": "^4.5.0",
     "webpack-dev-server": "^3.11.2"
   }
