+++ public/index.html
@@ -1,12 +1,13 @@
+<!DOCTYPE html>
 <html>
 
 <head>
-  <title>Hello world App</title>
-<script defer src="bundle.js"></script><script defer src="bundle.js"></script><script defer src="bundle.js"></script></head>
+  <title>Reviews</title>
+</head>
 
 <body>
   <div id="root"></div>
-  <script src="bundle.js"></script>
+  <script type="text/javascript" src="bundle.js"></script>
 </body>
 
 </html>
