+++ webpack.config.js
@@ -21,10 +21,6 @@ module.exports = {
         exclude: /node_modules/,
         // use the babel-loader for transpiling JavaScript to a suitable format
         loader: 'babel-loader',
-        options: {
-          // attach the presets to the loader (most projects use .babelrc file instead)
-          presets: ['@babel/preset-env', '@babel/preset-react']
-        }
       }
     ]
   },
